# src/main/java/com/epam/gym/dto/AuthenticationDTO.java
package com.epam.gym.dto;


import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class AuthenticationDTO {
	
	private String username;
	private String password;
}