# src/main/java/com/epam/gym/dto/ChangePasswordRequest.java
package com.epam.gym.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class ChangePasswordRequest {

	private String currentPassword;
	private String newPassword;
	private String confirmationPassword;
}